Made all logic for full search by categories and attributes and for reservation of the item

+++ Server_side/src/item/item.controller.ts
@@ -1,7 +1,7 @@
 import {
     Body,
     Controller,
-    Delete,
+    Delete, ForbiddenException,
     Get,
     Param,
     Post,
@@ -29,11 +29,43 @@ export class ItemController {
     @ApiQuery({ name: 'minPrice', required: false, type: Number })
     @ApiQuery({ name: 'maxPrice', required: false, type: Number })
     @ApiQuery({ name: 'sellerId', required: false, type: Number })
-    findAll(@Query() paginationDto: PaginationDto, @Query('typeId') typeId?: number, @Query('minPrice') minPrice?: number,
-            @Query('maxPrice') maxPrice?: number, @Query('sellerId') sellerId?: number) {
-        return this.itemService.findAll(paginationDto, { typeId, minPrice, maxPrice, sellerId });
+    @ApiQuery({ name: 'attributes', required: false, type: String, description: 'JSON string of attributes filter' })
+    findAll(
+        @Query() paginationDto: PaginationDto,
+        @Query('typeId') typeId?: number,
+        @Query('minPrice') minPrice?: number,
+        @Query('maxPrice') maxPrice?: number,
+        @Query('sellerId') sellerId?: number,
+        @Query('attributes') attributes?: string
+    ) {
+        const attributeFilters = attributes ? JSON.parse(attributes) : {};
+        return this.itemService.findAll(paginationDto, { typeId, minPrice, maxPrice, sellerId, attributes: attributeFilters });
     }
 
+    @Get('reserved')
+    @ApiResponse({ status: 200, description: 'Your reserved items has been successfully retrieved.'})
+    @ApiResponse({ status: 401, description: "You don't have permission to retrieve this!"})
+    @ApiResponse({ status: 403, description: "You don't have rights to retrieve this!"})
+    @UseGuards(JwtAuthGuard)
+    async getReservedItems(@Req() req) {
+        if (req.user.role !== 'buyer') {
+            return await this.itemService.getReservedItems(req.user.userId);
+        } else {
+            throw new ForbiddenException("You dont have rights to make or store reservations!");
+        }
+    }
+
+    @Get('pending-approval')
+    @ApiResponse({ status: 200, description: 'Items that are pending approval has been successfully retrieved.'})
+    @ApiResponse({ status: 401, description: "You don't have permission to retrieve this!"})
+    @ApiResponse({ status: 403, description: "You don't have rights to retrieve this!"})
+    @UseGuards(JwtAuthGuard)
+    async getItemsPendingApproval(@Req() req) {
+        if (req.user.role !== 'seller') {
+            throw new ForbiddenException("You don't have rights to watch pending approvals!");
+        }
+        return await this.itemService.getItemsPendingApproval(req.user.userId);
+    }
 
     @Get(':id')
     @ApiResponse({ status: 200, description: 'Item has been found.'})
@@ -55,13 +87,70 @@ export class ItemController {
             role: request.user.role,
         };
 
+        body.userId = user.userId;
+
         body.images = files?.map((file) => file.path) || []
 
         return this.itemService.create(body, user);
     }
 
+    @Post(':id/reserve')
+    @UseGuards(JwtAuthGuard)
+    @ApiResponse({ status: 200, description: 'Item successfully reserved.'})
+    @ApiResponse({ status: 401, description: 'Unauthorized access.'})
+    @ApiResponse({ status: 403, description: 'You don\'t have rights to reserve items!'})
+    @ApiResponse({ status: 404, description: 'Item not found.'})
+    async reserve(@Param('id') itemId: number, @Req() req) {
+        if (req.user.role === 'buyer') {
+            return await this.itemService.reserveItem(itemId, req.user.userId);
+        } else {
+            throw new ForbiddenException("You don't have rights to reserve items!");
+        }
+    }
+
+    @Delete(':id/reserve')
+    @UseGuards(JwtAuthGuard)
+    @ApiResponse({ status: 200, description: 'Item successfully removed from reserved list.'})
+    @ApiResponse({ status: 401, description: 'Unauthorized access.'})
+    @ApiResponse({ status: 403, description: 'You don\'t have rights to reserve items!'})
+    @ApiResponse({ status: 404, description: 'Item not found.'})
+    async deleteReservation(@Param('id') itemId: number, @Req() req) {
+        if (req.user.role === 'buyer') {
+            return await this.itemService.removeReservation(itemId, req.user.userId);
+        } else {
+            throw new ForbiddenException("You don't have rights to remove reserved items!");
+        }
+    }
+
+    @Post(':id/approve')
+    @ApiResponse({ status: 200, description: 'Reservation has been successfully approved.'})
+    @ApiResponse({ status: 401, description: "You don't have permission to approve this reservation!"})
+    @ApiResponse({ status: 403, description: "You don't have rights to approve this reservation!"})
+    @UseGuards(JwtAuthGuard)
+    async approveReservation( @Param('id') itemId: number, @Req() req) {
+        if (req.user.role !== 'seller') {
+            throw new ForbiddenException("You don't have rights to approve reservations!");
+        }
+        return await this.itemService.approveReservation(itemId, req.user.userId);
+    }
+
+    @Post(':id/reject')
+    @ApiResponse({ status: 200, description: 'Reservation has been successfully rejected.'})
+    @ApiResponse({ status: 401, description: "You don't have permission to reject this reservation!"})
+    @ApiResponse({ status: 403, description: "You don't have rights to reject this reservation!"})
+    @UseGuards(JwtAuthGuard)
+    async rejectReservation(@Param('id') itemId: number, @Req() req) {
+        if (req.user.role !== 'seller') {
+            throw new ForbiddenException("You don't have rights to reject reservations!");
+        }
+        return await this.itemService.rejectReservation(itemId, req.user.userId);
+    }
 
     @Put(':id')
+    @ApiResponse({ status: 200, description: 'Item has been successfully updated.'})
+    @ApiResponse({ status: 401, description: "You don't have permission to edit this item!"})
+    @ApiResponse({ status: 403, description: "You don't have rights to edit this item!"})
+    @ApiResponse({ status: 404, description: 'Item not found.'})
     @UseGuards(JwtAuthGuard)
     @UseInterceptors(FilesInterceptor('images'))
     async update(
@@ -70,6 +159,9 @@ export class ItemController {
         @UploadedFiles() files: Express.Multer.File[],
         @Req() request,
     ) {
+        if (request.user.role !== 'seller') {
+            throw new ForbiddenException("You don't have rights to edit items!");
+        }
         const existingImages = await this.itemService.retrieveExistingImages(id);
         const images = files ? files.map((file) => file.path) : [];
 
@@ -82,8 +174,13 @@ export class ItemController {
     @UseGuards(JwtAuthGuard)
     @ApiResponse({ status: 200, description: 'Item has been successfully deleted.'})
     @ApiResponse({ status: 401, description: "You don't have permission to delete this item!"})
+    @ApiResponse({ status: 403, description: "You don't have rights to delete this item!"})
     @ApiResponse({ status: 404, description: 'Item not found.'})
     delete(@Param('id') id: number, @Req() request) {
+        if (request.user.role !== 'seller') {
+            throw new ForbiddenException("You don't have rights to delete the item!");
+        }
+
         const userId = request.user.id;
 
         return this.itemService.delete(id, userId);
